Joi (#36)
* Début de schemas

* ++

* LIBRE !!!!!!!!!!!!!!!

+++ routes/dungeon/index.ts
@@ -2,6 +2,7 @@ import express from 'express';
 import pagesize from "../../queryConfig";
 import { DungeonRepository } from '../../src/interfaces/repository/dungeon.repository';
 import PrismaDungeonRepository from '../../src/implementation/repository/prisma/dungeon.repository';
+import { Schemas } from '../../schemas/index';
 
 const router = express.Router();
 // Should be DI
@@ -9,6 +10,11 @@ const repository : DungeonRepository = new PrismaDungeonRepository();
 
 // Get all dungeons
 router.get("/", async (req: any, res: any) => {
+    var { error, value } = Schemas.dungeonsGetSchema.validate(req.query);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     var {page, size} = req.query;
     if(page===undefined || page<=0){
         page=1;
@@ -18,10 +24,21 @@ router.get("/", async (req: any, res: any) => {
         res.status(400).send("Error: "+err);
         return;
     });
-    res.status(200).json(dungeons);
+    var { error, value } = Schemas.dungeonsSchema.validate(dungeons);
+    if (error === undefined) {
+        res.status(200).json(dungeons);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.get("/:id", async (req: any, res: any) => {
+    var { error, value } = Schemas.dungeonsGetByIdSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     const { id } = req.params;
     const dungeon = await repository.findById(parseInt(id)).catch((err)=>{
         res.status(400).send("Error: "+err);
@@ -31,10 +48,21 @@ router.get("/:id", async (req: any, res: any) => {
         res.status(404).send("Dungeon not found");
         return;
     }
-    res.status(200).json(dungeon);
+    var { error, value } = Schemas.dungeonSchema.validate(dungeon);
+    if (error === undefined) {
+        res.status(200).json(dungeon);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.get("/name/:name", async (req: any, res: any) => { 
+    var { error, value } = Schemas.dungeonsGetByNameSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     const { name } = req.params;
     const dungeon = await repository.findByName(name).catch((err)=>{
         res.status(400).send("Error: "+err);
@@ -44,10 +72,21 @@ router.get("/name/:name", async (req: any, res: any) => {
         res.status(404).send("Dungeon not found");
         return;
     }
-    res.status(200).json(dungeon);
+    var { error, value } = Schemas.dungeonSchema.validate(dungeon);
+    if (error === undefined) {
+        res.status(200).json(dungeon);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.post("/", async (req: any, res: any) => { 
+    var { error, value } = Schemas.dungeonsPostSchema.validate(req.body);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
     const { name } = req.body;
     // TODO : Validation
 
@@ -55,23 +94,47 @@ router.post("/", async (req: any, res: any) => {
         res.status(400).send("Error: "+err);
         return;
     });
-    res.status(201).json(dungeon);
+    var { error, value } = Schemas.dungeonSchema.validate(dungeon);
+    if (error === undefined) {
+        res.status(201).json(dungeon);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 });
 
 router.delete("/:id", async (req: any, res: any)=>{
+    var { error, value } = Schemas.dungeonsDeleteSchema.validate(req.params);
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
+
 	const {id}= req.params;
 	
 	const dungeon =await repository.delete(parseInt(id)).catch((err)=>{
         res.status(400).send("Error: "+err);
         return;
     });
-    res.status(201).json(dungeon);
+    var { error, value } = Schemas.dungeonSchema.validate(dungeon);
+    if (error === undefined) {
+        res.status(201).json(dungeon);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 	
 });
 
 router.put("/:id", async (req: any, res: any)=>{
 	const {id} = req.params;
     const {name}= req.body;
+
+    var { error, value } = Schemas.dungeonsPutSchema.validate({id, name});
+    if (error !== undefined) {
+        res.status(300).send("Error: "+ error);
+        return;
+    }
 
 	let notValid = !name;
     if(notValid){
@@ -83,7 +146,13 @@ router.put("/:id", async (req: any, res: any)=>{
         res.status(400).send("Error: "+err);
         return;
     });
-    res.status(201).json(dungeon);
+    var { error, value } = Schemas.dungeonSchema.validate(dungeon);
+    if (error === undefined) {
+        res.status(201).json(dungeon);
+    }
+    else {
+        res.status(500).send("Error: "+ error);
+    }
 	
 });
